Fix: build issues for deployment in vercel
