+++ app/contextProvider.ts
@@ -1,0 +1,7 @@
+import { createContext } from "react";
+import { TodoContextInterface } from "@/app/todoContext";
+
+export const TodoContext = createContext<TodoContextInterface>({
+  todoName: "",
+  setTodoName: () => {},
+});
