--- app/interface.ts
@@ -1,4 +1,0 @@
-export interface TodoContextInterface {
-    todoName: string;
-    setTodoName: Function;
-}
